+++ packages/fabrik_layout/lib/fabrik_layout.dart
@@ -2,4 +2,5 @@ export 'src/fabrik_layout.dart';
 export 'src/fabrik_layout_data.dart';
 export 'src/fabrik_breakpoints.dart';
 export 'src/fabrik_layout_type.dart';
+export 'src/fabrik_text_scale_config.dart';
 export 'src/context_extensions.dart';